gvariant: Fix reading a uint64 on Linux
guint64 is UInt there, not UInt64.

+++ Frida/GLib/Variant.swift
@@ -65,7 +65,7 @@ extension GLib {
         }
 
         public var uint64: UInt64? {
-            isOfType("t") ? g_variant_get_uint64(handle) : nil
+            isOfType("t") ? UInt64(g_variant_get_uint64(handle)) : nil
         }
 
         public var string: String? {
